１ファイルにまとめる。ただし重複があるので消す予定のものもある
