+++ src/features/JSONPath/index.ts
@@ -1,3 +1,2 @@
 export * from "./core";
-export * from "./pluginOld";
 export * from "./top";
