+++ src/features/JSONPath/plugin.test.ts
@@ -8,8 +8,11 @@ import type {
 } from "@RmmzPluginSchema/rmmz/plugin";
 import type { PluginCommandData } from "@RmmzPluginSchema/rmmz/plugin/types/pluginCommand";
 import { JSONPathJS } from "jsonpath-js";
-import { mergeCommandMap } from "./pluginOld";
-import { extractArgsFromPluiginCommand, jsonPathFromPluginSchema } from "./top";
+import {
+  extractArgsFromPluiginCommand,
+  jsonPathFromPluginSchema,
+  mergeCommandMap,
+} from "./top";
 
 const minimumSchema: PluginMinimumSchema = {
   pluginName: "MockPlugin",
