+++ src/features/JSONPath/top.buildPath.test.ts
@@ -2,16 +2,13 @@ import { describe, expect, test } from "vitest";
 import type { PluginReadResult } from "@RmmzPluginSchema/fileio/types";
 import type {
   ParsedPlugin,
-  PluginCommandData,
   PluginParam,
   PluginParamsRecord,
   PluginParamTokens,
 } from "@RmmzPluginSchema/rmmz/plugin";
 import { JSONPathJS } from "jsonpath-js";
 import type { ConvertPluginResult, PluginExtractedValue } from "./core";
-import { mergeCommandMap } from "./pluginOld";
 import {
-  extractArgsFromPluiginCommand,
   jsonPathFromPluginReadResult,
   jsonPathFromPluginReadResults,
 } from "./top";
@@ -174,43 +171,5 @@ describe("jsonPathFromPluginReadResults", () => {
     expect(result[1].schema.pluginName).toBe("PluginC");
     expect(result[0].params).toEqual(expectedParams0);
     expect(result[1].params).toEqual(expectedParams1);
-  });
-});
-
-describe("command extraction integration", () => {
-  test("JSONPath変換結果からコマンド抽出APIが新契約で動作する", () => {
-    const readResult = createReadResult("PluginA", "10", "Hero");
-
-    const convertResult = jsonPathFromPluginReadResult(readResult, (path) => {
-      return new JSONPathJS(path);
-    });
-
-    expect(convertResult).not.toBeNull();
-    const commandMap = mergeCommandMap([convertResult!]);
-
-    // 正常系: 定義済みコマンドを抽出
-    const command: PluginCommandData = {
-      code: 357,
-      parameters: ["PluginA", "Add", "Add", { value: "42" }],
-    };
-
-    const extracted = extractArgsFromPluiginCommand(command, commandMap);
-    expect(extracted.error).toBeUndefined();
-    expect(extracted.args).toHaveLength(1);
-    expect(extracted.args[0].value).toBe(42);
-
-    // 異常系: 未定義コマンドはerror付き結果を返す
-    const undefinedCommand: PluginCommandData = {
-      code: 357,
-      parameters: ["PluginA", "Unknown", "Unknown", {}],
-    };
-
-    const undefinedExtracted = extractArgsFromPluiginCommand(
-      undefinedCommand,
-      commandMap,
-    );
-    expect(undefinedExtracted.error).toBeDefined();
-    expect(undefinedExtracted.error?.source).toBe("undefinedCommand");
-    expect(undefinedExtracted.args).toEqual([]);
   });
 });
